satorilib/sqlite/oldstuff/requirements.txt
pandas
dask
distributed
sqlite3